agent/requirements.txt
pyyaml==6.0
requests==2.28.1
python-dotenv==0.19.0
